Time: 190 ms (15.29%), Space: 86.5 MB (6.47%) - LeetHub

+++ 1701-average-waiting-time/1701-average-waiting-time.cpp
@@ -14,6 +14,6 @@ public:
             }
         }
         
-        return double(totalWaitingTime) / customers.size();
+        return totalWaitingTime / customers.size();
     }
 };
